zad hani

+++ AISD/pracownie/B/reverse.cpp
@@ -9,7 +9,7 @@
 using namespace std;
 
 struct Point {
-  long long x,y;
+  int x,y;
 };
 
 double distance(Point a, Point b) {
@@ -39,9 +39,9 @@ bool compareY(Point a, Point b) {
 }
 
 double divAndConq(int n, const Point points[], const vector<Point> &pointsByY, array<Point, 3> &minArr) {
-  if(n < 3){
+  if(n < 3) {
     return DBL_MAX;
-  }  
+  }
 
   int mid = n/2;
 
@@ -117,19 +117,20 @@ int main() {
   array<Point, 3> min;
   points.reserve(n);
   for (int i = 0; i < n; i++) {
-    long long x,y; 
-    scanf("%lld %lld", &x, &y);
+    int x,y; 
+    scanf("%d %d", &x, &y);
     points.push_back({x,y});
   }
   
+
   sort(points.begin(), points.end(), compareX);
   vector<Point> pointsByY = points;
   sort(pointsByY.begin(), pointsByY.end(), compareY);
   
   divAndConq(points.size(), &points[0], pointsByY, min);
 
-  for(int i = 0; i < 3; ++i) {
-    printf("%lld %lld\n", min[i].x, min[i].y);
+  for(int i=0;i<3;++i) {
+    printf("%d %d\n", min[i].x, min[i].y);
   }
   
 }
